Ukrainian translation fixes
Notes for subsequent editors:

I see no point in using "uk_UA" since all other languages use the generic term ("en"/"fr"/"de"/"es") with the exception of literally a few cases when this is needed ("zh-Hans"/"zh-Hant").
This is not a case where adding "_UA" is mandatory, so it's better to use a generalized term.

Also, I am exploring the problem of using the terms "додаток", "програма", and "застосунок" as a translation for "application" term in our language, because there has been a lot of discussion and debate about this over time. Very controversial thing, requires some research.
If the use of the term "застосунок" in this case is the most correct, I will leave it unchanged.